requirements.txt
pytesseract
opencv-python-headless
pymupdf
pillow
numpy
matplotlib